refactor: ♻️ (`/[statLabel]`) グラフ部分を`<PopulationChart>`に分割して読み込み表示を追加した。  (#28)
